+++ internal/campaigns/campaign_spec.go
@@ -66,6 +66,7 @@ type Step struct {
 	Run       string            `json:"run,omitempty" yaml:"run"`
 	Container string            `json:"container,omitempty" yaml:"container"`
 	Env       map[string]string `json:"env,omitempty" yaml:"env"`
+	Files     map[string]string `json:"files,omitempty" yaml:"files,omitempty"`
 
 	image string
 }