+++ internal/campaigns/run_steps.go
@@ -10,6 +10,7 @@ import (
 	"os"
 	"os/exec"
 	"strings"
+	"text/template"
 	"time"
 
 	"github.com/hashicorp/go-multierror"
@@ -60,9 +61,15 @@ func runSteps(ctx context.Context, wc *WorkspaceCreator, repo *graphql.Repositor
 		return nil, errors.Wrap(err, "git commit failed")
 	}
 
+	results := make([]StepResult, len(steps))
+
 	for i, step := range steps {
 		logger.Logf("[Step %d] docker run %s %q", i+1, step.Container, step.Run)
-		reportProgress(step.Run)
+
+		stepContext := StepContext{Repository: *repo}
+		if i > 0 {
+			stepContext.PreviousStep = results[i-1]
+		}
 
 		cidFile, err := ioutil.TempFile(tempDir, repo.Slug()+"-container-id")
 		if err != nil {
@@ -87,29 +94,81 @@ func runSteps(ctx context.Context, wc *WorkspaceCreator, repo *graphql.Repositor
 
 		// Set up a temporary file on the host filesystem to contain the
 		// script.
-		fp, err := ioutil.TempFile(tempDir, "")
+		runScriptFile, err := ioutil.TempFile(tempDir, "")
 		if err != nil {
 			return nil, errors.Wrap(err, "creating temporary file")
 		}
-		hostTemp := fp.Name()
-		defer os.Remove(hostTemp)
-		if _, err := fp.WriteString(step.Run); err != nil {
-			return nil, errors.Wrapf(err, "writing to temporary file %q", hostTemp)
-		}
-		fp.Close()
+		defer os.Remove(runScriptFile.Name())
 
+		// Parse step.Run as a template...
+		tmpl, err := parseAsTemplate("step-run", step.Run, &stepContext)
+		if err != nil {
+			return nil, errors.Wrap(err, "parsing step run")
+		}
+
+		// ... and render it into a buffer and the temp file we just created.
+		var runScript bytes.Buffer
+		if err := tmpl.Execute(io.MultiWriter(&runScript, runScriptFile), stepContext); err != nil {
+			return nil, errors.Wrap(err, "executing template")
+		}
+		if err := runScriptFile.Close(); err != nil {
+			return nil, errors.Wrap(err, "closing temporary file")
+		}
+
+		// Parse and render the step.Files.
+		files, err := renderStepFiles(step.Files, &stepContext)
+		if err != nil {
+			return nil, errors.Wrap(err, "parsing step files")
+		}
+
+		// Create temp files with the rendered content of step.Files so that we
+		// can mount them into the container.
+		filesToMount := make(map[string]*os.File, len(files))
+		for name, content := range files {
+			fp, err := ioutil.TempFile(tempDir, "")
+			if err != nil {
+				return nil, errors.Wrap(err, "creating temporary file")
+			}
+			defer os.Remove(fp.Name())
+
+			if _, err := io.Copy(fp, content); err != nil {
+				return nil, errors.Wrap(err, "writing to temporary file")
+			}
+
+			if err := fp.Close(); err != nil {
+				return nil, errors.Wrap(err, "closing temporary file")
+			}
+
+			filesToMount[name] = fp
+		}
+
+		// Render the step.Env variables as templates.
+		env, err := renderStepEnv(step.Env, &stepContext)
+		if err != nil {
+			return nil, errors.Wrap(err, "parsing step files")
+		}
+
+		reportProgress(runScript.String())
 		const workDir = "/work"
-		cmd := exec.CommandContext(ctx, "docker", "run",
+		args := []string{
+			"run",
 			"--rm",
 			"--cidfile", cidFile.Name(),
 			"--workdir", workDir,
 			"--mount", fmt.Sprintf("type=bind,source=%s,target=%s", volumeDir, workDir),
-			"--mount", fmt.Sprintf("type=bind,source=%s,target=%s,ro", hostTemp, containerTemp),
-			"--entrypoint", shell,
-		)
-		for k, v := range step.Env {
-			cmd.Args = append(cmd.Args, "-e", k+"="+v)
+			"--mount", fmt.Sprintf("type=bind,source=%s,target=%s,ro", runScriptFile.Name(), containerTemp),
 		}
+		for target, source := range filesToMount {
+			args = append(args, "--mount", fmt.Sprintf("type=bind,source=%s,target=%s,ro", source.Name(), target))
+		}
+
+		for k, v := range env {
+			args = append(args, "-e", k+"="+v)
+		}
+
+		args = append(args, "--entrypoint", shell)
+
+		cmd := exec.CommandContext(ctx, "docker", args...)
 		cmd.Args = append(cmd.Args, "--", step.image, containerTemp)
 		cmd.Dir = volumeDir
 
@@ -132,7 +191,7 @@ func runSteps(ctx context.Context, wc *WorkspaceCreator, repo *graphql.Repositor
 			return nil, stepFailedErr{
 				Err:         err,
 				Args:        cmd.Args,
-				Run:         step.Run,
+				Run:         runScript.String(),
 				Container:   step.Container,
 				TmpFilename: containerTemp,
 				Stdout:      strings.TrimSpace(stdoutBuffer.String()),
@@ -141,10 +200,22 @@ func runSteps(ctx context.Context, wc *WorkspaceCreator, repo *graphql.Repositor
 		}
 
 		logger.Logf("[Step %d] complete in %s", i+1, elapsed)
-	}
 
-	if _, err := runGitCmd("add", "--all"); err != nil {
-		return nil, errors.Wrap(err, "git add failed")
+		if _, err := runGitCmd("add", "--all"); err != nil {
+			return nil, errors.Wrap(err, "git add failed")
+		}
+
+		statusOut, err := runGitCmd("status", "--porcelain")
+		if err != nil {
+			return nil, errors.Wrap(err, "git status failed")
+		}
+
+		changes, err := parseGitStatus(statusOut)
+		if err != nil {
+			return nil, errors.Wrap(err, "parsing git status output")
+		}
+
+		results[i] = StepResult{Files: changes, Stdout: &stdoutBuffer, Stderr: &stderrBuffer}
 	}
 
 	reportProgress("Calculating diff")
@@ -270,4 +341,167 @@ func (e stepFailedErr) SingleLineError() string {
 	}
 
 	return strings.Split(out, "\n")[0]
+}
+
+func parseAsTemplate(name, input string, stepCtx *StepContext) (*template.Template, error) {
+	return template.New(name).Delims("${{", "}}").Funcs(stepCtx.ToFuncMap()).Parse(input)
+}
+
+func renderStepFiles(files map[string]string, stepCtx *StepContext) (map[string]io.Reader, error) {
+	containerFiles := make(map[string]io.Reader, len(files))
+
+	for fileName, fileRaw := range files {
+		// We treat the file contents as a template and render it
+		// into a buffer that we then mount into the code host.
+		var out bytes.Buffer
+
+		tmpl, err := parseAsTemplate(fileName, fileRaw, stepCtx)
+		if err != nil {
+			return containerFiles, err
+		}
+
+		if err := tmpl.Execute(&out, stepCtx); err != nil {
+			return containerFiles, err
+		}
+
+		containerFiles[fileName] = &out
+	}
+
+	return containerFiles, nil
+}
+
+func renderStepEnv(env map[string]string, stepCtx *StepContext) (map[string]string, error) {
+	parsedEnv := make(map[string]string, len(env))
+
+	fnMap := stepCtx.ToFuncMap()
+
+	for k, v := range env {
+		// We treat the file contents as a template and render it
+		// into a buffer that we then mount into the code host.
+		var out bytes.Buffer
+
+		tmpl, err := template.New(k).Delims("${{", "}}").Funcs(fnMap).Parse(v)
+		if err != nil {
+			return parsedEnv, err
+		}
+
+		if err := tmpl.Execute(&out, stepCtx); err != nil {
+			return parsedEnv, err
+		}
+
+		parsedEnv[k] = out.String()
+	}
+
+	return parsedEnv, nil
+}
+
+// StepContext represents the contextual information available when executing a
+// step that's defined in a campaign spec.
+type StepContext struct {
+	PreviousStep StepResult
+	Repository   graphql.Repository
+}
+
+// ToFuncMap returns a template.FuncMap to access fields on the StepContext in a
+// text/template.
+func (stepCtx *StepContext) ToFuncMap() template.FuncMap {
+	return template.FuncMap{
+		"join": func(list []string, sep string) string {
+			return strings.Join(list, sep)
+		},
+		"split": func(s string, sep string) []string {
+			return strings.Split(s, sep)
+		},
+		"previous_step": func() map[string]interface{} {
+			result := map[string]interface{}{
+				"modified_files": stepCtx.PreviousStep.ModifiedFiles(),
+				"added_files":    stepCtx.PreviousStep.AddedFiles(),
+				"deleted_files":  stepCtx.PreviousStep.DeletedFiles(),
+				"renamed_files":  stepCtx.PreviousStep.RenamedFiles(),
+			}
+
+			if stepCtx.PreviousStep.Stdout != nil {
+				result["stdout"] = stepCtx.PreviousStep.Stdout.String()
+			} else {
+				result["stdout"] = ""
+			}
+
+			if stepCtx.PreviousStep.Stderr != nil {
+				result["stderr"] = stepCtx.PreviousStep.Stderr.String()
+			} else {
+				result["stderr"] = ""
+			}
+
+			return result
+		},
+		"repository": func() map[string]interface{} {
+			return map[string]interface{}{
+				"search_result_paths": stepCtx.Repository.SearchResultPaths(),
+				"name":                stepCtx.Repository.Name,
+			}
+		},
+	}
+}
+
+// StepResult represents the result of a previously executed step.
+type StepResult struct {
+	// Files are the changes made to files by the step.
+	Files StepChanges
+
+	// Stdout is the output produced by the step on standard out.
+	Stdout *bytes.Buffer
+	// Stderr is the output produced by the step on standard error.
+	Stderr *bytes.Buffer
+}
+
+// StepChanges are the changes made to files by a previous step in a repository.
+type StepChanges struct {
+	Modified []string
+	Added    []string
+	Deleted  []string
+	Renamed  []string
+}
+
+// ModifiedFiles returns the files modified by a step.
+func (r StepResult) ModifiedFiles() []string { return r.Files.Modified }
+
+// AddedFiles returns the files added by a step.
+func (r StepResult) AddedFiles() []string { return r.Files.Added }
+
+// DeletedFiles returns the files deleted by a step.
+func (r StepResult) DeletedFiles() []string { return r.Files.Deleted }
+
+// RenamedFiles returns the new name of files that have been renamed by a step.
+func (r StepResult) RenamedFiles() []string { return r.Files.Renamed }
+
+func parseGitStatus(out []byte) (StepChanges, error) {
+	result := StepChanges{}
+
+	stripped := strings.TrimSpace(string(out))
+	if len(stripped) == 0 {
+		return result, nil
+	}
+
+	for _, line := range strings.Split(stripped, "\n") {
+		if len(line) < 4 {
+			return result, fmt.Errorf("git status line has unrecognized format: %q", line)
+		}
+
+		file := line[3:]
+
+		switch line[0] {
+		case 'M':
+			result.Modified = append(result.Modified, file)
+		case 'A':
+			result.Added = append(result.Added, file)
+		case 'D':
+			result.Deleted = append(result.Deleted, file)
+		case 'R':
+			files := strings.Split(file, " -> ")
+			newFile := files[len(files)-1]
+			result.Renamed = append(result.Renamed, newFile)
+		}
+	}
+
+	return result, nil
 }